requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
Pillow>=8.2.0
numpy>=1.20.0
facenet-pytorch=2.5.2
matplotlib>=3.4.2
tqdm>=4.61.0